adding user and auth  model

+++ controllers/feed.js
@@ -1,5 +1,6 @@
 const { validationResult } = require("express-validator");
 const Post = require("../models/post");
+const User = require("../models/user");
 const fs = require("fs");
 const path = require("path");
 
@@ -9,6 +10,7 @@ exports.getPosts = (req, res, next) => {
   let totalItems;
   Post.find()
     .countDocuments()
+    .populate("creator")
     .then((count) => {
       totalItems = count;
       return Post.find()
@@ -47,19 +49,29 @@ exports.createPosts = (req, res, next) => {
 
   const imageUrl = req.file.path.replace("\\", "/");
   const { title, content } = req.body;
+  let creator;
 
   const post = new Post({
     title,
     content,
     imageUrl,
-    creator: { name: "MaxCool" },
+    creator: req.userId,
   });
   post
     .save()
     .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      creator = user;
+      user.posts.push(post);
+      return user.save();
+    })
+    .then((result) => {
       res.status(201).json({
         message: "Post created successfully",
-        post: result,
+        post,
+        creator: { _id: creator._id, name: creator.name },
       });
     })
     .catch((err) => {
@@ -123,6 +135,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorize");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -161,10 +178,22 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not Authorize");
+        error.statusCode = 403;
+        throw error;
+      }
       // check login user
       clearImage(post.imageUrl);
 
       return Post.findByIdAndRemove(postId);
+    })
+    .then((result) => {
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      user.posts.pull(postId);
+      return user.save();
     })
     .then((result) => {
       res.status(200).json({
